Updated map.html
The JS for creating and centering the map. Now the map item is created and zoom is set separately. It then attempts to use the users position through the navigator section of the Google maps API: if allowed it will set the center of the map to the user's location, if not it will default to look at campus.