Displaying stats and balancing
I have adjusted the random opponent method and stats so that a random opponent is gotten excluding your profile, it will then check that they have stats if not give them 1, after it will use a factor to raise or lower you opponents stats to be more similar to yours before loading the page.

This is then loaded in a simple bootstrap table designed to fit on either desktop or mobile.