Change Code and Scan to use random Code
This alters the event_id for the code and scan pages to now be the random 75 character code generated when the event is created. This code is generated from capital letters using the crypto library so will be extremely random. The QR code generated also stores this code and Scan now works for it.