Updated model to ass points and pictures
Updated models to add for spendable points which are only an int variable atm. Profile pics can also be specified for a user. The pics shall be stored in a media folder that have created. If you want to folder moved somewhere else you can but you should change to MEDIA URL and MEDIA ROOT variables. The profile tab displays the profile picture. Default picture is the "you" icon.