Fixed error for no opponents
I have fixed the error for when there are no opponents in the battle system.

I have added a check for if there are no other users of the system, if so then it will create a new CustomUser called Skeleton warrior to use with random stats of 1 to 3. This is not saved so is never in the DB.